// data/models/SkillCategory.ts
enum SkillCategory {
	BACKEND = 'backend',
	CLOUD = 'cloud',
	DATABASES = 'databases',
	DEVOPS = 'devops',
	FRAMEWORKS = 'frameworks',
	FRONTEND = 'frontend',
	LANGUAGES = 'languages',
	LIBRARIES = 'libraries',
	SOFTWARE_ARCHITECTURE = 'software-architecture',
	SOFT_SKILLS = 'soft-skills',
	TESTING = 'testing',
	TOOLS = 'tools',
}

export default SkillCategory;